+++ packages/colony-js-client/src/ColonyClient/senders/DomainAuth.js
@@ -183,7 +183,10 @@ export default class DomainAuth<
             ),
           );
 
-          const highestDomain = await this.getHighestDomain(permissionDomains);
+          const highestDomain = await this.getHighestDomain(
+            permissionDomains,
+            role,
+          );
 
           const childSkillIndexes = await Promise.all(
             domainIds.map(async domainId =>
@@ -204,7 +207,10 @@ export default class DomainAuth<
     return proofs.reduce((acc, proof) => ({ ...acc, ...proof }));
   }
 
-  async getHighestDomain(domainIds: number[]): Promise<number> {
+  async getHighestDomain(
+    domainIds: number[],
+    role: ColonyRole,
+  ): Promise<number> {
     const skills = await Promise.all(
       domainIds.map(async domainId => {
         if (domainId < 0) return [domainId, null];
@@ -220,12 +226,20 @@ export default class DomainAuth<
         return [domainId, skill];
       }),
     );
-    return skills
+    const filteredSkills = skills
       .filter(([, skill]) => skill !== null)
       .sort(
         // $FlowFixMe these will all be skills because of the filter
         ([, a], [, b]) => b.children.length - a.children.length,
-      )[0][0];
+      );
+    if (!filteredSkills.length) {
+      throw new Error(
+        `No permission for role ${role} in appropriate domain(s) for method ${
+          this.name
+        }`,
+      );
+    }
+    return filteredSkills[0][0];
   }
 
   async getChildSkillIndex(
@@ -237,7 +251,7 @@ export default class DomainAuth<
       throw new Error('Current wallet address does not have permission');
     } else if (permissionDomainId === childDomainId) {
       // if we have permission in immediate domain we know index will be zero
-      return 0; // TODO: is this correct?
+      return 0;
     }
 
     // get the permission and child domain skills