+++ packages/colony-js-client/src/__tests__/ColonyClient.test.js
@@ -17,6 +17,7 @@ const colonyEvents = [
   'ColonyFundsMovedBetweenFundingPots',
   'ColonyInitialised',
   'ColonyRewardInverseSet',
+  'ColonyRoleSet',
   'ColonyRootRoleSet',
   'ColonyUpgraded',
   'DomainAdded',