+++ packages/colony-js-contract-loader-network/src/NetworkLoader.js
@@ -8,7 +8,6 @@ import ContractLoader, {
 import type { RequiredContractProps } from '@colony/colony-js-contract-loader';
 
 const NETWORKS = {
-  RINKEBY: 'rinkeby',
   GOERLI: 'goerli',
 };
 
@@ -16,28 +15,26 @@ const DEFAULT_NETWORK = NETWORKS.GOERLI;
 
 type Network = $Values<typeof NETWORKS>;
 
-const LATEST_VERSION = 1;
+const LATEST_VERSION = 2;
 
 const CONTRACTS_MANIFEST = {
-  // static: [],
   versioned: {
-    [NETWORKS.RINKEBY]: {
-      '1': ['Authority', 'EtherRouter', 'IColony', 'IColonyNetwork', 'Token'],
-      '2': ['Authority', 'EtherRouter', 'IColony', 'IColonyNetwork', 'Token'],
-      '3': [
+    [NETWORKS.GOERLI]: {
+      '1': [
         'EtherRouter',
         'IColony',
         'IColonyNetwork',
         'IMetaColony',
         'IRecovery',
-        'Token',
-        'DSToken',
         'ITokenLocking',
+        'OldRoles',
+        'OldRolesFactory',
         'OneTxPayment',
+        'OneTxPaymentFactory',
+        'Token',
+        'TokenAuthority',
       ],
-    },
-    [NETWORKS.GOERLI]: {
-      '1': [
+      '2': [
         'EtherRouter',
         'IColony',
         'IColonyNetwork',
@@ -54,15 +51,6 @@ const CONTRACTS_MANIFEST = {
     },
   },
 };
-
-// const STATIC_CONTRACTS = CONTRACTS_MANIFEST.static.reduce(
-//   (contracts, contractName) =>
-//     Object.assign(contracts, {
-//       // eslint-disable-next-line global-require, import/no-dynamic-require
-//       [contractName]: require(`../contracts/static/${contractName}.json`),
-//     }),
-//   {},
-// );
 
 const VERSIONED_CONTRACTS = Object.entries(CONTRACTS_MANIFEST.versioned).reduce(
   (networks, [network, versions]) =>
@@ -105,14 +93,6 @@ class NetworkLoader extends ContractLoader {
 
     assert(!!contractName, 'A `contractName` option must be provided');
     assert(!!version, 'A valid `version` option must be provided');
-
-    // if (STATIC_CONTRACTS[contractName]) {
-    //   return this._transform(
-    //     STATIC_CONTRACTS[contractName],
-    //     networkQuery,
-    //     requiredProps,
-    //   );
-    // }
 
     const contract =
       VERSIONED_CONTRACTS[network] &&