+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 
 * Fix `DomainAuth` sender to account for `defaultValues` (`@colony/colony-js-client`)
 * Fix `makeExecuteTaskRoleChange` input to use `address` (`@colony/colony-js-client`)
+* Patch `lookupRegisteredENSDomain` for testnets. Domains were previously being returned ending in `.eth` on networks where the deployed ENS uses the `.test` TLD (`@colony/colony-js-client`)
+* Fix `getLogs` method to correctly format topic filters (`@colony/colony-js-contract-client`)
+
+**New Features**
+
+* Added `goerli` version 2 deployment to network loader (`@colony/colony-js-contract-loader-network`)
+
+**Maintenance**
+
+* Add `ColonyRoleSet` event, which replaces the existing role set events as the output of any methods which previously used them. The existing events remain in the client for the purpose of parsing logs which were created by previous versions of the contract (`@colony/colony-js-client`)
+* Improve thrown error for `DomainAuth` sender methods when the current address does not have permission (`@colony/colony-js-client`)
+* Remove `rinkeby` deployments from network loader, as they are no longer supported (`@colony/colony-js-contract-loader-network`)
+* Remove unused network support from `getNetworkClient` (`@colony/colony-js-client`)
 
 ## v1.12.1